+++ src/Codeception/Module/Phalcon1.php
@@ -7,6 +7,9 @@ use Phalcon\Di;
 use Phalcon\DiInterface;
 use Phalcon\Di\Injectable;
 use Phalcon\Mvc\Model as PhalconModel;
+use Phalcon\Mvc\Url;
+use Phalcon\Mvc\Router\RouteInterface;
+use Phalcon\Mvc\RouterInterface;
 use Codeception\TestCase;
 use Codeception\Configuration;
 use Codeception\Lib\Connector\Phalcon as PhalconConnector;
@@ -14,6 +17,7 @@ use Codeception\Lib\Framework;
 use Codeception\Lib\Interfaces\ActiveRecord;
 use Codeception\Lib\Interfaces\PartedModule;
 use Codeception\Exception\ModuleConfigException;
+use Exception;
 use Codeception\Lib\Connector\PhalconMemorySession;
 
 /**
@@ -27,7 +31,7 @@ use Codeception\Lib\Connector\PhalconMemorySession;
  * ## Status
  *
  * * Maintainer: **Serghei Iakovlev**
- * * Stability: **dev**
+ * * Stability: **stable**
  * * Contact: sadhooklay@gmail.com
  *
  * ## Example
@@ -143,6 +147,7 @@ class Phalcon1 extends Framework implements ActiveRecord, PartedModule
         Di::setDefault($this->di);
 
         if ($this->di->has('session')) {
+            // Destroy existing sessions of previous tests
             $this->di['session'] = new PhalconMemorySession();
         }
 
@@ -208,6 +213,17 @@ class Phalcon1 extends Framework implements ActiveRecord, PartedModule
     }
 
     /**
+     * Provides access the Phalcon application object.
+     *
+     * @see \Codeception\Lib\Connector\Phalcon::getApplication
+     * @return \Phalcon\Mvc\Application|\Phalcon\Mvc\Micro|\Phalcon\Cli\Console
+     */
+    public function getApplication()
+    {
+        $this->client->getApplication();
+    }
+
+    /**
      * Sets value to session. Use for authorization.
      *
      * @param string $key
@@ -215,25 +231,66 @@ class Phalcon1 extends Framework implements ActiveRecord, PartedModule
      */
     public function haveInSession($key, $val)
     {
-        $this->di->get('session')->set($key, (string)$val);
-        $this->debugSection('Session', json_encode($this->di['session']->getAll()));
+        $this->di->get('session')->set($key, $val);
+        $this->debugSection('Session', json_encode($this->di['session']->toArray()));
     }
 
     /**
      * Checks that session contains value.
      * If value is `null` checks that session has key.
      *
+     * ``` php
+     * <?php
+     * $I->seeInSession('key');
+     * $I->seeInSession('key', 'value');
+     * ?>
+     * ```
+     *
      * @param string $key
      * @param mixed $value
      */
     public function seeInSession($key, $value = null)
     {
-        $this->debugSection('Session', json_encode($this->di['session']->getAll()));
-        if (is_null($value)) {
-            $this->assertTrue($this->di['session']->has($key));
+        $this->debugSection('Session', json_encode($this->di['session']->toArray()));
+
+        if (is_array($key)) {
+            $this->seeSessionHasValues($key);
             return;
         }
-        $this->assertEquals($value, $this->di['session']->get($key));
+
+        if (!$this->di['session']->has($key)) {
+            $this->fail("No session variable with key '$key'");
+        }
+
+        if (is_null($value)) {
+            $this->assertTrue($this->di['session']->has($key));
+        } else {
+            $this->assertEquals($value, $this->di['session']->get($key));
+        }
+    }
+
+    /**
+     * Assert that the session has a given list of values.
+     *
+     * ``` php
+     * <?php
+     * $I->seeSessionHasValues(['key1', 'key2']);
+     * $I->seeSessionHasValues(['key1' => 'value1', 'key2' => 'value2']);
+     * ?>
+     * ```
+     *
+     * @param  array $bindings
+     * @return void
+     */
+    public function seeSessionHasValues(array $bindings)
+    {
+        foreach ($bindings as $key => $value) {
+            if (is_int($key)) {
+                $this->seeInSession($value);
+            } else {
+                $this->seeInSession($key, $value);
+            }
+        }
     }
 
     /**
@@ -381,6 +438,60 @@ class Phalcon1 extends Framework implements ActiveRecord, PartedModule
             $this->fail($e->getMessage());
 
             return null;
+        }
+    }
+
+    /**
+     * Opens web page using route name and parameters.
+     *
+     * ``` php
+     * <?php
+     * $I->amOnRoute('posts.create');
+     * ?>
+     * ```
+     *
+     * @param $routeName
+     * @param array $params
+     */
+    public function amOnRoute($routeName, $params = [])
+    {
+        if (!$this->di->has('url')) {
+            $this->fail('Unable to resolve "url" service.');
+        }
+
+        /** @var Url $url */
+        $url = $this->di->getShared('url');
+
+        try {
+            $this->amOnPage($url->get(['for' => $routeName], null, true));
+        } catch (Exception $e) {
+            $this->fail($e->getMessage());
+        }
+    }
+
+    /**
+     * Checks that current url matches route
+     *
+     * ``` php
+     * <?php
+     * $I->seeCurrentRouteIs('posts.index');
+     * ?>
+     * ```
+     * @param string $routeName
+     */
+    public function seeCurrentRouteIs($routeName)
+    {
+        if (!$this->di->has('url')) {
+            $this->fail('Unable to resolve "url" service.');
+        }
+
+        /** @var Url $url */
+        $url = $this->di->getShared('url');
+
+        try {
+            $this->seeCurrentUrlEquals($url->get(['for' => $routeName], null, true));
+        } catch (Exception $e) {
+            $this->fail($e->getMessage());
         }
     }