+++ src/Codeception/Module/REST.php
@@ -18,10 +18,9 @@ use Codeception\Util\Soap as XmlUtils;
  * Module for testing REST WebService.
  *
  * This module can be used either with frameworks or PHPBrowser.
- * It tries to guess the framework is is attached to.
- *
- * Whether framework is used it operates via standard framework modules.
- * Otherwise sends raw HTTP requests to url via PHPBrowser.
+ * If a framework module is connected, the testing will occur in the application directly.
+ * Otherwise, a PHPBrowser should be specified as a dependency to send requests and receive responses from a server.
+ * 
  *
  * ## Status
  *
@@ -81,6 +80,10 @@ EOF;
      */
     public $client = null;
     public $isFunctional = false;
+
+    /**
+     * @var InnerBrowser
+     */
     protected $connectionModule;
 
     public $headers = [];
@@ -137,7 +140,7 @@ EOF;
 
     private function getRunningClient()
     {
-        if ($this->client->getHistory()->isEmpty()) {
+        if ($this->client->getInternalRequest() === null) {
             throw new ModuleException($this, "Response is empty. Use `\$I->sendXXX()` methods to send HTTP request");
         }
         return $this->client;
@@ -475,12 +478,11 @@ EOF;
                 $this->debugSection("Request", "$method $url " . json_encode($parameters));
             }
             $this->client->request($method, $url, $parameters, $files);
-
         } else {
             $this->debugSection("Request", "$method $url " . $parameters);
             $this->client->request($method, $url, [], $files, [], $parameters);
         }
-        $this->response = (string)$this->client->getInternalResponse()->getContent();
+        $this->response = (string)$this->connectionModule->_getResponseContent();
         $this->debugSection("Response", $this->response);
 
         if (count($this->client->getInternalRequest()->getCookies())) {
@@ -516,7 +518,7 @@ EOF;
      */
     public function seeResponseIsJson()
     {
-        json_decode($this->response);
+        json_decode($this->connectionModule->_getResponseContent());
         $errorCode = json_last_error();
         \PHPUnit_Framework_Assert::assertEquals(
             JSON_ERROR_NONE,
@@ -534,7 +536,7 @@ EOF;
      */
     public function seeResponseContains($text)
     {
-        $this->assertContains($text, $this->response, "REST response contains");
+        $this->assertContains($text, $this->connectionModule->_getResponseContent(), "REST response contains");
     }
 
     /**
@@ -546,7 +548,7 @@ EOF;
      */
     public function dontSeeResponseContains($text)
     {
-        $this->assertNotContains($text, $this->response, "REST response contains");
+        $this->assertNotContains($text, $this->connectionModule->_getResponseContent(), "REST response contains");
     }
 
     /**
@@ -575,7 +577,7 @@ EOF;
      */
     public function seeResponseContainsJson($json = [])
     {
-        $jsonResponseArray = new JsonArray($this->response);
+        $jsonResponseArray = new JsonArray($this->connectionModule->_getResponseContent());
         \PHPUnit_Framework_Assert::assertTrue(
             $jsonResponseArray->containsArray($json),
             "Response JSON contains provided\n"
@@ -603,7 +605,7 @@ EOF;
      */
     public function grabResponse()
     {
-        return $this->response;
+        return $this->connectionModule->_getResponseContent();
     }
 
     /**
@@ -631,7 +633,7 @@ EOF;
      */
     public function grabDataFromResponseByJsonPath($jsonPath)
     {
-        return (new JsonArray($this->response))->filterByJsonPath($jsonPath);
+        return (new JsonArray($this->connectionModule->_getResponseContent()))->filterByJsonPath($jsonPath);
     }
 
     /**
@@ -676,9 +678,10 @@ EOF;
      */
     public function seeResponseJsonMatchesXpath($xpath)
     {
+        $response = $this->connectionModule->_getResponseContent();
         $this->assertGreaterThan(
-            0, (new JsonArray($this->response))->filterByXPath($xpath)->length,
-            "Received JSON did not match the XPath `$xpath`.\nJson Response: \n" . $this->response
+            0, (new JsonArray($response))->filterByXPath($xpath)->length,
+            "Received JSON did not match the XPath `$xpath`.\nJson Response: \n" . $response
         );
     }
 
@@ -727,9 +730,10 @@ EOF;
      */
     public function seeResponseJsonMatchesJsonPath($jsonPath)
     {
+        $response = $this->connectionModule->_getResponseContent();
         $this->assertNotEmpty(
-            (new JsonArray($this->response))->filterByJsonPath($jsonPath),
-            "Received JSON did not match the JsonPath provided\n" . $this->response
+            (new JsonArray($response))->filterByJsonPath($jsonPath),
+            "Received JSON did not match the JsonPath provided\n" . $response
         );
     }
 
@@ -741,9 +745,10 @@ EOF;
      */
     public function dontSeeResponseJsonMatchesJsonPath($jsonPath)
     {
+        $response = $this->connectionModule->_getResponseContent();
         $this->assertEmpty(
-            (new JsonArray($this->response))->filterByJsonPath($jsonPath),
-            "Received JSON did (but should not) match the JsonPath provided\n" . $this->response
+            (new JsonArray($response))->filterByJsonPath($jsonPath),
+            "Received JSON did (but should not) match the JsonPath provided\n" . $response
         );
     }
 
@@ -755,7 +760,7 @@ EOF;
      */
     public function dontSeeResponseContainsJson($json = [])
     {
-        $jsonResponseArray = new JsonArray($this->response);
+        $jsonResponseArray = new JsonArray($this->connectionModule->_getResponseContent());
         $this->assertFalse(
             $jsonResponseArray->containsArray($json),
             "Response JSON does not contain JSON provided\n"
@@ -845,7 +850,7 @@ EOF;
      */
     public function seeResponseMatchesJsonType(array $jsonType, $jsonPath = null)
     {
-        $jsonArray = new JsonArray($this->response);
+        $jsonArray = new JsonArray($this->connectionModule->_getResponseContent());
         if ($jsonPath) {
             $jsonArray = $jsonArray->filterByJsonPath($jsonPath);
         }
@@ -864,7 +869,7 @@ EOF;
      */
     public function dontSeeResponseMatchesJsonType($jsonType, $jsonPath = null)
     {
-        $jsonArray = new JsonArray($this->response);
+        $jsonArray = new JsonArray($this->connectionModule->_getResponseContent());
         if ($jsonPath) {
             $jsonArray = $jsonArray->filterByJsonPath($jsonPath);
         }
@@ -879,9 +884,9 @@ EOF;
      * @part xml
      * @param $response
      */
-    public function seeResponseEquals($response)
+    public function seeResponseEquals($expected)
     {
-        $this->assertEquals($response, $this->response);
+        $this->assertEquals($expected, $this->connectionModule->_getResponseContent());
     }
 
     /**
@@ -917,7 +922,7 @@ EOF;
     public function seeResponseIsXml()
     {
         libxml_use_internal_errors(true);
-        $doc = simplexml_load_string($this->response);
+        $doc = simplexml_load_string($this->connectionModule->_getResponseContent());
         $num = "";
         $title = "";
         if ($doc === false) {
@@ -946,7 +951,7 @@ EOF;
      */
     public function seeXmlResponseMatchesXpath($xpath)
     {
-        $structure = new XmlStructure($this->response);
+        $structure = new XmlStructure($this->connectionModule->_getResponseContent());
         $this->assertTrue($structure->matchesXpath($xpath), 'xpath not matched');
     }
 
@@ -962,7 +967,7 @@ EOF;
      */
     public function dontSeeXmlResponseMatchesXpath($xpath)
     {
-        $structure = new XmlStructure($this->response);
+        $structure = new XmlStructure($this->connectionModule->_getResponseContent());
         $this->assertTrue($structure->matchesXpath($xpath), 'accidentally matched xpath');
     }
 
@@ -976,7 +981,7 @@ EOF;
      */
     public function grabTextContentFromXmlElement($cssOrXPath)
     {
-        $el = (new XmlStructure($this->response))->matchElement($cssOrXPath);
+        $el = (new XmlStructure($this->connectionModule->_getResponseContent()))->matchElement($cssOrXPath);
         return $el->textContent;
     }
 
@@ -991,7 +996,7 @@ EOF;
      */
     public function grabAttributeFrom($cssOrXPath, $attribute)
     {
-        $el = (new XmlStructure($this->response))->matchElement($cssOrXPath);
+        $el = (new XmlStructure($this->connectionModule->_getResponseContent()))->matchElement($cssOrXPath);
         if (!$el->hasAttribute($attribute)) {
             $this->fail("Attribute not found in element matched by '$cssOrXPath'");
         }
@@ -1009,7 +1014,7 @@ EOF;
      */
     public function seeXmlResponseEquals($xml)
     {
-        \PHPUnit_Framework_Assert::assertXmlStringEqualsXmlString($this->response, $xml);
+        \PHPUnit_Framework_Assert::assertXmlStringEqualsXmlString($this->connectionModule->_getResponseContent(), $xml);
     }
 
 
@@ -1024,7 +1029,7 @@ EOF;
      */
     public function dontSeeXmlResponseEquals($xml)
     {
-        \PHPUnit_Framework_Assert::assertXmlStringNotEqualsXmlString($this->response, $xml);
+        \PHPUnit_Framework_Assert::assertXmlStringNotEqualsXmlString($this->connectionModule->_getResponseContent(), $xml);
     }
 
     /**
@@ -1044,7 +1049,7 @@ EOF;
      */
     public function seeXmlResponseIncludes($xml)
     {
-        $this->assertContains(XmlUtils::toXml($xml)->C14N(), XmlUtils::toXml($this->response)->C14N(), "found in XML Response");
+        $this->assertContains(XmlUtils::toXml($xml)->C14N(), XmlUtils::toXml($this->connectionModule->_getResponseContent())->C14N(), "found in XML Response");
     }
 
     /**
@@ -1057,7 +1062,7 @@ EOF;
      */
     public function dontSeeXmlResponseIncludes($xml)
     {
-        $this->assertNotContains(XmlUtils::toXml($xml)->C14N(), XmlUtils::toXml($this->response)->C14N(), "found in XML Response");
+        $this->assertNotContains(XmlUtils::toXml($xml)->C14N(), XmlUtils::toXml($this->connectionModule->_getResponseContent())->C14N(), "found in XML Response");
     }
 
     /**