+++ tests/unit/Codeception/Module/TestsForWeb.php
@@ -61,6 +61,14 @@ abstract class TestsForWeb extends \PHPUnit_Framework_TestCase
         $this->module->dontSee('Welcome','h6');
     }
 
+    public function testSeeInSource()
+    {
+        $this->module->amOnPage('/');
+        $this->module->seeInSource('<h1>Welcome to test app!</h1>');
+        $this->module->seeInSource('A wise man said: "debug!"');
+        $this->module->dontSeeInSource('John Cleese');
+    }
+
     public function testSeeInCurrentUrl()
     {
         $this->module->amOnPage('/info');