+++ tests/unit/Codeception/Util/JsonTypeTest.php
@@ -7,14 +7,16 @@ class JsonTypeTest extends \Codeception\TestCase\Test
         'id' => 'integer:>10',
         'retweeted' => 'Boolean',
         'in_reply_to_screen_name' => 'null|string',
+        'name' => 'string|null', // http://codeception.com/docs/modules/REST#seeResponseMatchesJsonType
         'user' => [
-          'url' => 'String:url'
+            'url' => 'String:url'
         ]
     ];
     protected $data = [
         'id' => 11,
         'retweeted' => false,
         'in_reply_to_screen_name' => null,
+        'name' => null,
         'user' => ['url' => 'http://davert.com']
     ];